requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=0.21.0
sqlalchemy>=1.4.41
psycopg2>=2.9.6
pydantic>=1.10.7
passlib[bcrypt]>=1.7.4